requirements_for_slackbot.txt
Flask==2.0.1
python-dotenv==0.19.0
slackclient==2.9.3
slackeventsapi==3.0.0
